Make guicast/stringfile.C handle longer length
Fixed a bug in the calculation, by including the indentation when
figuring out if there is room enough, and dropped an unused variable.

+++ guicast/stringfile.h
@@ -30,7 +30,7 @@
 class StringFile
 {
 public:
-	StringFile(long length = 0);
+	StringFile(size_t length = 0);
 	StringFile(char *filename);
 	virtual ~StringFile();
 
@@ -53,13 +53,13 @@ public:
 	int writeline(char *arg1, Freq arg2, int indent);   // write next line to string
 	int backupline();       // move back one line
 
-	long get_length();
-	long get_pointer();
+	size_t get_length();
+	size_t get_pointer();
 	int write_to_file(char *filename);
 	int read_from_string(char *string);
 
 	char *string;
-	long pointer, length, available;
+	size_t pointer, length, available;
 	char string1[1024];      // general purpose strings
 };
